Add TimelockController.sol to resolve TCC-01, DTC-04, TCT-06, TVT-01
This adds a standard implementation of OpenZepplin's TimelockController to the project in order to resolve the centralisation issues (TCC-01, DTC-04, TCT-06, TVT-01) discussed in Certik's audit of TipsyCoin. Two TimelockControllers will be used for the deployment (one for ProxyAdmin, one for Owner), both with a 48 hour delay. The Proposers for these TimelockController's will be The 3/5 Gnosis Multisig for ProxyAdmin, and the 2/3 Gnosis Multisig for the Owner. The Executors for these TimelockControllers will be the members of those respective Multisigs, plus the Deployment wallet. This info, plus additional commentary can be found in the updated README.md